+++ packages/auto/changelog/src/write-changelog-files.ts
@@ -1,4 +1,4 @@
-import type { THook, TPackageRelease, TMessage, TLogReleaseType } from '@auto/core'
+import type { THook, TPackageRelease, TMessage, TLogReleaseType } from '@nextnextools/auto-core'
 import { isString } from 'tsfn'
 import type { TReadonly } from 'tsfn'
 import { writeChangelog } from './write-changelog'