+++ packages/auto/commit-prompt/src/types.ts
@@ -1,4 +1,4 @@
-import type { TAutoConfig, TRequiredPrefixes } from '@auto/core'
+import type { TAutoConfig, TRequiredPrefixes } from '@nextnextools/auto-core'
 
 export type TPrefixesWithTitles = {
   [k: string]: {