+++ packages/auto/slack/test/send-slack-message.ts
@@ -1,4 +1,4 @@
-import type { TPackageRelease, TAutoConfig } from '@auto/core'
+import type { TPackageRelease, TAutoConfig } from '@nextnextools/auto-core'
 import { mockRequire } from '@mock/require'
 import { createSpy, getSpyCalls } from 'spyfn'
 import test from 'tape'